feat(render): light the day, follow the hero, give bodies weight
Five coordinated changes to the three.js 2D compositor, designed in parallel
and applied in sequence so each built on a verified tree.

Terrain and lighting. district-lighting.ts keyed on MapId alone, so a district
looked identical at 3am and noon while atmosphere.ts already read absoluteMinute.
worldSun() now interpolates a six-key table across the 1440-minute day: sun
direction, colour temperature, and shadow length and direction sweep
continuously. The four keys at the old bucket centres reproduce today's shadow
vectors exactly. Worst per-minute step is 0.18px, midnight wrap included,
against the old table's 21px seams. Sprite shadows rotate at fixed length, so
technique 6's measured value survives at every minute. Ground gains a per-corner
light field through the existing vertex tint attribute; measured T-junction
error is 1.09 of 255. Batches, materials, programs and textures all unchanged.

Camera. Follow with a dead zone and frame-rate-independent smoothing, trauma
shake bounded to the 4 world px / 180 ms envelope docs/specs requires, and a
pure sampleCameraDirector queue. clampCamera now snaps to whole screen pixels
rather than whole world pixels; the old rounding froze follow ~1.5 world px
short of target at any zoom above 1. Five consumers assumed an integer camera,
including one inside the verify gate. Shake lives only in renderCamera and never
in persisted state, and the preference debounce now skips while shots are
queued so a scripted composition can never reach disk.

Motion. walkFrame flips every 32 world pixels of travel, which makes
fract(travelDistance / 32) an exact stride clock that is already deterministic.
The bob dips into the down pose while the knee absorbs and peaks at passing,
because a naive sine reads as a bounce rather than weight. Turn lean, stop lean,
foot-plant events and impact poses are all pure functions of state the
simulation already produces. No engine, no run state, no domain change.

VFX. Click ripples, ground dust, footstep dust, and a violence kit driven by a
stub combat interface, plus a scripted shooting scene in the dev harness. Every
cap is quoted from docs/specs/2026-08-11-skia-procedural-vfx.md rather than
invented. Peak 13 transient rects, 14% of the budget, +1 draw call.

UI. Reanimated is installed but imported nowhere and gets no UI thread on web,
so motion is raw CSS instead: an opacity-only state layer, panel entry
animations, and an elevation scale. role="button" on 16 controls that exposed
as generic. #world-audio-caption had no zIndex under a 50-layer conversation
overlay, so the caption for players who cannot hear vocal cues was occluded
every time it fired. Dead ContextActionMenu.tsx deleted.

Docs. CLAUDE.md called src/render a "Skia world scene"; Skia was removed in
stage 7. It also omitted three from the layer-purity list despite
import-boundaries.ts having a dedicated check, and listed six evidence nodes
where there are now nine.

Verified: typecheck, build:electron, check:boundaries, verify:first-hour, and
104 suites / 1071 tests, up from 97 / 956 with no regressions. Reviewed by
Claude Fable 5 at xhigh, which independently recomputed the sun curve and traced
every integer-camera consumer; verdict approved, three should-fix items applied.

Not verified: no packaged smoke ran. The packaged app dies on this machine with
GPU exit_code=15 and a macOS Mach port permission failure. Sun sweep, muzzle
flash, batch counts and the button fix were confirmed in a browser via the web
export instead. The gated ground-1:1 change is not included; it needs pixels.

+++ scripts/electron/run-responsive-package-smoke.ts
@@ -52,7 +52,9 @@ const EvidenceSchema = z.object({
   automaticWorldZoom: z.union([z.literal(1), z.literal(2), z.literal(3)]),
   selectedWorldZoom: z.union([z.literal(1), z.literal(2), z.literal(3)]),
   uiScale: z.union([z.literal(1), z.literal(1.25), z.literal(1.5)]),
-  camera: z.object({ x: z.number().int(), y: z.number().int(), zoom: z.union([z.literal(1), z.literal(2), z.literal(3)]) }).strict(),
+  // The camera snaps to a whole screen pixel, not a whole world pixel, so x and y are fractional
+  // at zoom 2 and 3. See clampCamera in src/render/camera.ts.
+  camera: z.object({ x: z.number(), y: z.number(), zoom: z.union([z.literal(1), z.literal(2), z.literal(3)]) }).strict(),
   mapId: z.literal('northwest_residential'),
   artMode: z.enum(['legacy', 'enhanced']),
   presentationHash: z.string().regex(/^[0-9a-f]{8}$/u),
